+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 - Added `@@categorized-childs-manage` that will render a link to manage
   categorized elements.
   [gbastien]
+- Implemented `CategorizedContent._unrestrictedGetObject` as `collective.eeafaceted.z3ctable`
+  calls now `item._unrestrictedGetObject` instead `item.getObject` for performance optimization.
+  [gbastien]
 
 0.53 (2022-04-22)
 -----------------